requirements.txt
pandas
sqlalchemy
mysql-connector-python
openpyxl
python-dotenv